# Conflicts:
#	_maps/Prefab/Departments.dmm
#	_maps/map_files/BoxStation/BoxStation.dmm
#	_maps/map_files/CorgStation/CorgStation.dmm
#	_maps/map_files/Deltastation/DeltaStation2.dmm
#	_maps/map_files/EchoStation/EchoStation.dmm
#	_maps/map_files/FlandStation/FlandStation.dmm
#	_maps/map_files/KiloStation/KiloStation.dmm
#	_maps/map_files/MetaStation/MetaStation.dmm
#	_maps/map_files/RadStation/RadStation.dmm
#	_maps/map_files/generic/CentCom.dmm
#	_maps/shuttles/mining/mining_rad.dmm
#	beestation.dme
#	code/__DEFINES/directional.dm
#	code/__DEFINES/radio.dm
#	code/game/machinery/requests_console.dm

+++ code/game/objects/items/implants/implantuplink.dm
@@ -9,9 +9,14 @@
 	/// The uplink flags of the implant uplink inside, only checked during initialisation so modifying it after initialisation will do nothing
 	var/uplink_flag = UPLINK_TRAITORS
 
+CREATION_TEST_IGNORE_SUBTYPES(/obj/item/implant/uplink)
+
 /obj/item/implant/uplink/Initialize(mapload, owner, uplink_flag)
 	. = ..()
-	AddComponent(/datum/component/uplink, _owner = owner, _lockable = TRUE, _enabled = FALSE, uplink_flag = uplink_flag, starting_tc = starting_tc)
+	if(!uplink_flag)
+		uplink_flag = src.uplink_flag
+	var/datum/component/uplink/new_uplink = AddComponent(/datum/component/uplink, _owner = owner, _lockable = TRUE, _enabled = FALSE, uplink_flag = uplink_flag, starting_tc = starting_tc)
+	new_uplink.unlock_text = "Your Syndicate Uplink has been cunningly implanted in you, for a small TC fee. Simply trigger the uplink to access it."
 	RegisterSignal(src, COMSIG_COMPONENT_REMOVING, PROC_REF(_component_removal))
 
 /**
@@ -22,12 +27,17 @@
  * the component, so delete itself.
  */
 /obj/item/implant/uplink/proc/_component_removal(datum/source, datum/component/component)
+	SIGNAL_HANDLER
+	if(QDELETED(src))
+		return
 	if(istype(component, /datum/component/uplink))
 		qdel(src)
 
 /obj/item/implanter/uplink
 	name = "implanter (uplink)"
 	imp_type = /obj/item/implant/uplink
+
+CREATION_TEST_IGNORE_SUBTYPES(/obj/item/implanter/uplink)
 
 /obj/item/implanter/uplink/Initialize(mapload, uplink_flag = UPLINK_TRAITORS)
 	imp = new imp_type(src, null, uplink_flag)
